[TG-19][TG-20] feat: confirmar/rebutjar incidencia (vots)
S'ha creat la logica d'actualitzar incidencia al eliminar/crear vots.
Implementar el calcul de l'score de la data en funcio de quants dies han passat des del report.
S'ha posat una nova columna data_score

Nota:
Falta utilitar la fiabilitat de l'usuari
Falta obligar que el accepted sigui -1 o 1 a Vote.java

+++ src/main/java/com/safesteps/backend/domain/incidents/dto/VoteResponseDTO.java
@@ -12,7 +12,7 @@ public class VoteResponseDTO {
     private  Long id;
     private Long incidenceId;
     private Long userId;
-    private float score;
+    private double score;
     private LocalDateTime createdAt;
 
     public VoteResponseDTO(Vote v) {
